refactor:라우터분리,디렉토리컨벤션 수정
