+++ ai_server/__init__.py
@@ -1,4 +1,4 @@
 """
 AI Text Transformation Server
-__version__ = '2.0.0'
+__version__ = '2.1.0'
 """
