+++ ai_server/api/__init__.py
@@ -1,0 +1,1 @@
+# API 패키지
