+++ ai_server/api/v1/comments.py
@@ -1,0 +1,48 @@
+from fastapi import APIRouter, HTTPException
+from ai_server.schemas.converter_schemas import CommentRequest, CommentResponse
+from ai_server.service.cat import cat_converter
+from ai_server.service.dog import dog_converter
+
+router = APIRouter()
+
+@router.post("/", 
+    response_model=CommentResponse,
+    responses={
+        200: {"model": CommentResponse, "description": "Successfully transformed text"},
+        400: {"model": CommentResponse, "description": "Empty Input"},
+        422: {"model": CommentResponse, "description": "wrong post_type"},
+        500: {"model": CommentResponse, "description": "internal_server_error"}
+    }
+)
+async def generate_comment(request: CommentRequest):
+    """댓글 텍스트를 고양이/강아지 말투로 변환합니다."""
+    # 빈 입력 체크
+    if not request.content.strip():
+        raise HTTPException(status_code=400, detail="Empty Input")
+
+    try:
+        # post_type에 따라 적절한 변환기 선택하게
+        if request.post_type == "cat":
+            transformed_content = cat_converter(request.content)
+        elif request.post_type == "dog":
+            transformed_content = dog_converter(request.content)
+        else:
+            raise HTTPException(status_code=422, detail="wrong post_type")
+        
+        # 성공 응답
+        return CommentResponse(
+            status_code=200,
+            message="Successfully transformed text",
+            data=transformed_content
+        )
+        
+    except ValueError as ve:
+        # Enum 값이 잘못된 경우 등
+        if "Enum" in str(ve) or "post_type" in str(ve):
+            raise HTTPException(status_code=422, detail="wrong post_type")
+        else:
+            raise
+
+    except Exception:
+        # 기타 모든 에러는 500으로 처리
+        raise HTTPException(status_code=500, detail="internal_server_error")
