+++ ai_server/main.py
@@ -1,23 +1,21 @@
-from fastapi import FastAPI, Request, HTTPException
+from fastapi import FastAPI, Request
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
 from fastapi.exceptions import RequestValidationError
-from ai_server.post.post_schemas import PostRequest, PostResponse
-from ai_server.rule_based_converter.cat import cat_converter
-from ai_server.rule_based_converter.dog import dog_converter
-from ai_server.rule_based_converter.converter_schemas import CommentRequest, CommentResponse
 from starlette.exceptions import HTTPException as StarletteHTTPException
-from ai_server.post.post_model import PostTransformationService
+from ai_server.api.v1.api import api_router
 import logging
 
 # 로깅 설정
+logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 # FastAPI 앱 초기화
 app = FastAPI(
     title="AI Text Transformation Server",
     description="SNS 포스팅/댓글/채팅을 고양이/강아지 말투로 변환하는 AI API 서버",
-    version="0.1.0"
+    version="2.1.0",
+    docs_url="/docs",
 )
 
 # CORS 설정
@@ -32,7 +30,7 @@ app.add_middleware(
 # 검증 오류 처리기
 @app.exception_handler(RequestValidationError)
 async def validation_exception_handler(request: Request, exc: RequestValidationError):
-    # Enum 값 검증 오류 메시지 생성
+    """검증 오류 처리"""
     error_msg = "잘못된 입력값"
     
     # 오류 상세 정보 추출
@@ -56,6 +54,7 @@ async def validation_exception_handler(request: Request, exc: RequestValidationE
 # HTTP 예외 처리기
 @app.exception_handler(StarletteHTTPException)
 async def http_exception_handler(request: Request, exc: StarletteHTTPException):
+    """HTTP 예외 처리"""
     return JSONResponse(
         status_code=exc.status_code,
         content={
@@ -77,91 +76,21 @@ async def generic_exception_handler(request: Request, exc: Exception):
         }
     )
 
+# API 라우터 포함
+app.include_router(api_router)
+
 # 루트 엔드포인트
 @app.get("/")
 async def root():
-    return {"message": "SNS 포스팅/댓글/채팅을 고양이/강아지 말투로 변환하는 AI API 서버"}
-
-# 텍스트 변환 엔드포인트
-@app.post("/generate/post", 
-    response_model=PostResponse,
-    responses={
-        200: {"model": PostResponse, "description": "Successfully transformed text"},
-        400: {"model": PostResponse, "description": "Empty Input"},
-        422: {"model": PostResponse, "description": "wrong post_type or emotion"},
-        500: {"model": PostResponse, "description": "internal_server_error"}
+    """루트 엔드포인트 - 서버 상태 확인"""
+    return {
+        "message": "SNS 포스팅/댓글/채팅을 고양이/강아지 말투로 변환하는 AI API 서버",
+        "version": "2.1.0",
+        "status": "healthy"
     }
-)
-async def generate_post(request: PostRequest):
-    # 빈 입력 체크
-    if not request.content.strip():
-        raise HTTPException(status_code=400, detail="Empty Input")
 
-    try:
-        # 스키마에서 정의된 타입을 사용하여 포스트 변환 서비스 실행
-        post_service = PostTransformationService()  # 기본 모델 사용
-        transformed_content = await post_service.transform_post(
-            content=request.content,
-            emotion=request.emotion,
-            post_type=request.post_type
-        )
-        
-        # 성공 응답
-        return PostResponse(
-            status_code=200,
-            message="Successfully transformed text",
-            data=transformed_content
-        )
-        
-    except ValueError as ve:
-        # Enum 값이 잘못된 경우 등
-        if "Enum" in str(ve) or "post_type" in str(ve) or "emotion" in str(ve):
-            raise HTTPException(status_code=422, detail="wrong post_type or emotion")
-        else:
-            raise
-
-    except Exception:
-        # 기타 모든 에러는 500으로 처리
-        raise HTTPException(status_code=500, detail="internal_server_error")
-
-# 댓글 변환 엔드포인트
-@app.post("/generate/comment", 
-    response_model=CommentResponse,
-    responses={
-        200: {"model": CommentResponse, "description": "Successfully transformed text"},
-        400: {"model": CommentResponse, "description": "Empty Input"},
-        422: {"model": CommentResponse, "description": "wrong post_type"},
-        500: {"model": CommentResponse, "description": "internal_server_error"}
-    }
-)
-async def generate_comment(request: CommentRequest):
-    # 빈 입력 체크
-    if not request.content.strip():
-        raise HTTPException(status_code=400, detail="Empty Input")
-
-    try:
-        # post_type에 따라 적절한 변환기 선택하게
-        if request.post_type == "cat":
-            transformed_content = cat_converter(request.content)
-        elif request.post_type == "dog":
-            transformed_content = dog_converter(request.content)
-        else:
-            raise HTTPException(status_code=422, detail="wrong post_type")
-        
-        # 성공 응답
-        return CommentResponse(
-            status_code=200,
-            message="Successfully transformed text",
-            data=transformed_content
-        )
-        
-    except ValueError as ve:
-        # Enum 값이 잘못된 경우 등
-        if "Enum" in str(ve) or "post_type" in str(ve):
-            raise HTTPException(status_code=422, detail="wrong post_type")
-        else:
-            raise
-
-    except Exception:
-        # 기타 모든 에러는 500으로 처리
-        raise HTTPException(status_code=500, detail="internal_server_error")
+# 헬스체크 엔드포인트
+@app.get("/health")
+async def health_check():
+    """헬스체크 엔드포인트"""
+    return {"status": "healthy"}
