--- ai_server/rule_based_converter/__init__.py
@@ -1,15 +1,0 @@
-"""
-Rule-based text converter package for transforming text to cat/dog speech patterns.
-"""
-
-from .cat import cat_converter
-from .dog import dog_converter
-from .converter_schemas import CommentRequest, CommentResponse, CommentType
-
-__all__ = [
-    "cat_converter",
-    "dog_converter", 
-    "CommentRequest",
-    "CommentResponse",
-    "CommentType"
-]
