+++ ai_server/service/dog.py
